OrganizedFolders
Reads file representing a folder hierarchy & verifies that all entries & subentries are in alphabetical order
Assumes no lines are empty
If 2 lines in same subentry are not in alphabetical order, print message to screen saying so & print the 2 lines that violate the order

A machine contains folders & sub-folders nested to several levels. Sub-folders are indicated by deeper indentation. All the sub-folders of a given entry are preceded by the same number of space; that number is greater than the indentation at the previous level.

For example:

class
     abstract
     accessors
     constructors
          overloaded
          no-args
     modifiers
method
     private
     public
     static
stack
     for handling nested structures
     methods
          push
          top
          peekTop
zeta